pyproject.toml
[tool.poetry]
name = "homework-24-1"
version = "0.1.0"
description = ""
authors = ["angelina <angln.pnn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
psycopg2 = "^2.9.9"
pillow = "^10.2.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
redis = "^5.0.1"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
